17th Commit - Adicionando Transferência com Cartão de Crédito & Pagamento de Fatura

+++ src/main/java/com/lorian/lorianBank/Cartao/CartaoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface CartaoRepository extends JpaRepository<Cartao, Integer>{
+public interface CartaoRepository extends JpaRepository<Cartao, Long>{
 	Optional<Cartao> findByNumero(String numero);
 	void deleteByNumero(String numero);
 }
